FIX: Ensure brew is in PATH early enough.
The specific issue here was that the `asdf` plugin for oh-my-zsh checks if `brew` is in the PATH. If it is, it sources it from the install directory, if not, it does nothing.

Since we only set `brew` into PATH later on when we source our own startup scripts, this is too late in this case. The simple fix for now is to use the `brew` plugin before any other plugins, but this also exports some unwanted aliases.