Sway: Setup `swayidle` and `swaylock`
Swaylock: This will lock the screen and prompt for a password. It's possible to trigger it manually with the command alone. It requires the user password to unlock the screen again.

Swayidle: Is started through the sway config. It will then execute the timeout commands after X seconds IDLE. The executed commands can be anything, currently mostly using the example.