Navi: Fix ZSH widget with zsh-vi-mode.
As far as I understand, the "zsh-vi-mode" plugin builds on top of the ZLE builtin VI mode and simple redefined most functions/widgets to work better/support more features. In doing so, it also sets the input mode to VI.

This seems to cause an issue with the default Navi widget (https://github.com/denisidoro/navi/issues/328). Looking at the output of `bindkey` after sourcing the file, the function set for CTRL-G isn't overwritten by Navi. I assume this is because Navis `bindkey` command does not specify the keymap.

To fix this, we explicitly define the keymap command again, this time specifically for the `viins` keymap. This works as expected then.