WezTerm: New terminal to replace Alacritty.
This is a very minor upgrade from Alacritty, mainly because it supports ligatures.

I have disabled the whole default keymap, copied it into my config and only picked the parts that I think are needed. The rest of the config makes the terminal as minimal as possible, much like Alacritty.