Difftastic: Git symantic diff tool.
Configured as the default for `git diff` for now. I'll try if this works out well for me. It takes a bit longer to load and shows no whitespace changes (as expected).

There are now bail-out versions for diff that use delta. Also, adding patches uses delta for now. We'll see if I buy in more.