Helix: Added language configuration.
Until recently, this wasn't portable enough. But now I've added Lexical to the `$PATH` and can refer to it implicitly in the config file. This is required for this LSP specifically, because it must be built from source and the absolute path to the start script will be different on each system.

To add it to the path, I'm using a simple ZShell setup script under `~/.zsh_setup.d/20-lexical.sh` that exports the `$PATH`. Crucially, this file is not checked into this repository, because it isn't portable enough to work on other systems.

The only problematic part is that Helix does NOT show an error dialog if the path to the LSP can't be found - it just simply doesn't start the LSP.