Wezterm: Only use SUPER key modifier on macOS.
On Linux and Windows, shortcuts of a program are usually defined with the CTRL modifier instead. The shortcuts with the SUPER modifier are usually more global systemwide shortcuts.

This already clashes in my current config, for example with Sway which uses SUPER for its shortcuts and doesn't forward them to Wezterm anymore.